requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
sseclient-py==1.7
urllib3==1.22
